daemon: fold command-line options into the runtime context
Remove the standalone opts module (opts.h, opts.c) in the daemon, which
maintained a separate static global options object. Options are moved to
bf_ctx, and so are accessors.
