+++ src/bpfilter/cgen/stub.c
@@ -30,8 +30,8 @@
 #include "cgen/printer.h"
 #include "cgen/program.h"
 #include "cgen/swich.h"
+#include "ctx.h"
 #include "filter.h"
-#include "opts.h"
 
 #define _BF_LOW_EH_BITMASK 0x1801800000000801ULL
 
@@ -73,7 +73,7 @@ static int _bf_stub_make_ctx_dynptr(struct bf_program *program, int arg_reg,
              BPF_MOV32_IMM(BPF_REG_3, bf_program_error_counter_idx(program)));
         EMIT_FIXUP_ELFSTUB(program, BF_ELFSTUB_UPDATE_COUNTERS);
 
-        if (bf_opts_is_verbose(BF_VERBOSE_BPF))
+        if (bf_ctx_is_verbose(BF_VERBOSE_BPF))
             EMIT_PRINT(program, "failed to create a new dynamic pointer");
 
         r = program->runtime.ops->get_verdict(BF_VERDICT_ACCEPT);
@@ -132,7 +132,7 @@ int bf_stub_parse_l2_ethhdr(struct bf_program *program)
              BPF_MOV32_IMM(BPF_REG_3, bf_program_error_counter_idx(program)));
         EMIT_FIXUP_ELFSTUB(program, BF_ELFSTUB_UPDATE_COUNTERS);
 
-        if (bf_opts_is_verbose(BF_VERBOSE_BPF))
+        if (bf_ctx_is_verbose(BF_VERBOSE_BPF))
             EMIT_PRINT(program, "failed to create L2 dynamic pointer slice");
 
         r = program->runtime.ops->get_verdict(BF_VERDICT_ACCEPT);
@@ -207,7 +207,7 @@ int bf_stub_parse_l3_hdr(struct bf_program *program)
              BPF_MOV32_IMM(BPF_REG_3, bf_program_error_counter_idx(program)));
         EMIT_FIXUP_ELFSTUB(program, BF_ELFSTUB_UPDATE_COUNTERS);
 
-        if (bf_opts_is_verbose(BF_VERBOSE_BPF))
+        if (bf_ctx_is_verbose(BF_VERBOSE_BPF))
             EMIT_PRINT(program, "failed to create L3 dynamic pointer slice");
 
         r = program->runtime.ops->get_verdict(BF_VERDICT_ACCEPT);
@@ -366,7 +366,7 @@ int bf_stub_parse_l4_hdr(struct bf_program *program)
              BPF_MOV32_IMM(BPF_REG_3, bf_program_error_counter_idx(program)));
         EMIT_FIXUP_ELFSTUB(program, BF_ELFSTUB_UPDATE_COUNTERS);
 
-        if (bf_opts_is_verbose(BF_VERBOSE_BPF))
+        if (bf_ctx_is_verbose(BF_VERBOSE_BPF))
             EMIT_PRINT(program, "failed to create L4 dynamic pointer slice");
 
         r = program->runtime.ops->get_verdict(BF_VERDICT_ACCEPT);
